requirements.txt
# =============================================================================
# REQUIREMENTS.TXT - Actualizado para Python 3.13 (Entorno 2026)
# =============================================================================

# --- CORE DATA SCIENCE ---
pandas>=2.2.0
numpy>=2.0.0
scipy>=1.14.0

# --- VISUALIZACIÓN ---
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.24.0
kaleido>=0.2.1

# --- ESTADÍSTICA Y ML ---
statsmodels>=0.14.4
scikit-learn>=1.5.0

# --- DEEP LEARNING (Versiones compatibles con Python 3.13+) ---
tensorflow>=2.18.0
keras>=3.5.0

# --- NLP (FinBERT) ---
# Usamos el extra-index para asegurar la versión CPU de Torch
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.5.0
transformers>=4.45.0
tokenizers>=0.20.0
nltk>=3.9.0
wordcloud>=1.9.3

# --- NOTEBOOK RUNTIME (Python 3.13) ---
ipykernel>=6.29.0
ipython>=8.26.0
nbformat>=5.10.0
nbclient>=0.10.0
nbconvert>=7.16.0
ipywidgets>=8.1.2
jupyter_client>=8.6.0
jupyter_core>=5.7.2
jupyter>=1.0.0

# --- UTILIDADES Y SISTEMA ---
python-dotenv>=1.0.0
tqdm>=4.66.0
python-dateutil>=2.9.0
requests>=2.32.3
urllib3>=2.2.0